Create programmer_strings.cpp
A programmer string contains letters that can be rearranged to form the word 'programmer' and is a substring of a longer string. Note that the strings 'programmer', 'grammproer', and 'prozmerqgram' are all classified as programmer strings by definition. Given a string, determine the number of indices that lie between the rightmost and leftmost programmer strings that it contains.